pyproject.toml
[project]
name = "finny"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128",
    "agno[pgvector,postgres]>=2.3",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.0",
    "asyncpg>=0.29",
    "bcrypt>=4.0",
    "pyjwt>=2.8",
    "alembic>=1.13",
    "psycopg[binary]>=3.3.2",
]

[dependency-groups]
dev = ["ruff>=0.8", "pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27", "aiosqlite>=0.20", "greenlet>=3.0"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "UP046",  # generic class uses Generic subclass (still valid syntax)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
